Use VS Code as EDITOR if running within a VS Code terminal
When the shell is running within the built-in terminal in VS Code, use it as the text editor. Otherwise, use nano.

`--wait` holds onto the process used to open the file until you close the file's tab. This helps with tools (like git) which expect your changes to be done when the editor closes.